automatically scale fonts based on monitor DPI
Each time the urxvt window is moved, check to see if the window is now
on a monitor with a different DPI.  If so, calculate and set a new font
size so that the physical size of the text is unchanged.  Works great
for laptops with HiDPI displays and external monitors; may require some
tweaking for other scenarios (e.g. laptop + presentation projector).

Scaling can be disabled by setting ENABLE_DPI_SCALING to 0 in the
extension source.

Note: This feature requires the xrandr command-line tool to be
installed.