Fix font scaling on rotated outputs
When an output is rotated with xrandr, the command-line utility prints
the pixel dimensions of the display as rotated, but the physical
dimensions of the display without rotation.  As a result, when the
rotation is "left" or "right", they didn't match each other, and the
wrong scale was calculated.  Check explicitly for this.